requirements.txt
faiss-cpu
numpy
sentence-transformers
transformers
torch
fastapi
uvicorn
pydantic
pytest
httpx
